Basic calendar display done
